Fix database migration dariah_core for new install, READ NOTE PLEASE.
In a fresh install of version 1.0 the migrations would fail as it would
try renaming existing tables which don't exist on a fresh install (this
was due to app renaming).
This new migration will simply create the right tables in the database.

If you are doing a fresh install without any prior
existing database you will have no problems.

However, if you already had a database in place
do the following:
  1. Install version 1.0 and migrate dariah_core
  (it will rename the existing tables);
  2. Manually remove the (last) migration from the
  South migration table;
  3. Install version 1.0-1;
  4. Do a fake migrate for dariah_core
  ``manage.py migrate dariah_core 0001 --fake``.

Sorry for the inconvenience! Renaming a Django app
is not a trivial task. More info here:
http://www.sjkingston.com/blog/renaming-a-django-app-that-uses-south/

I opted for this route as there was only me and
the server that had a pre-existing database.